Updated to add easy, medium and hard
The code has been updated to adjust neural network weightings so that a player can't keep playing the same moves if they find a way to win. As this makes it very hard to play easy, medium and hard options have been added. Easy and Medium make occasional random moves (i.e. mistakes) to give the opponent a chance.